+++ src/main/java/codesquad/global/dao/ArticleQuery.java
@@ -1,7 +1,7 @@
 package codesquad.global.dao;
 
 import codesquad.article.domain.vo.Status;
-import codesquad.common.PageRequest;
+import codesquad.common.http.request.PageRequest;
 
 import java.util.List;
 import java.util.Optional;
